[BE] refactor : 폴더 구조 도메인형으로 변경
