--- backend/driving-today/src/main/java/com/drivingtoday/Entity/Academy.java
@@ -1,34 +1,0 @@
-package com.drivingtoday.Entity;
-
-
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
-
-@Builder
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class Academy {
-
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id",updatable = false)
-    private Long id;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "latitude")
-    private Double latitude;
-
-    @Column(name = "longitude")
-    private Double longitude;
-
-    @Column(name = "cert")
-    private Boolean cert;
-
-
-}
