--- backend/driving-today/src/main/java/com/drivingtoday/Entity/Instructor.java
@@ -1,41 +1,0 @@
-package com.drivingtoday.Entity;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
-
-@Builder
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class Instructor {
-
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id",updatable = false)
-    private Long id;
-
-    @Column(name = "name")
-    private String name;
-
-    @Column(name = "phone_number")
-    private String phone_number;
-
-    @Column(name = "instructor_image")
-    private String instructor_image;
-
-    @Column(name = "price_per_hour")
-    private String price_per_hour;
-
-    @Column(name = "introduction")
-    private String introduction;
-
-
-
-}
