--- backend/driving-today/src/main/java/com/drivingtoday/Entity/Reservation.java
@@ -1,35 +1,0 @@
-package com.drivingtoday.Entity;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
-
-@Builder
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class Reservation {
-
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id",updatable = false)
-    private Long id;
-
-    @Column(name = "status")
-    private String status;
-
-
-
-
-
-
-
-
-
-}
