--- backend/driving-today/src/main/java/com/drivingtoday/Entity/Student.java
@@ -1,38 +1,0 @@
-package com.drivingtoday.Entity;
-
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.NoArgsConstructor;
-import lombok.RequiredArgsConstructor;
-
-@Builder
-@Entity
-@NoArgsConstructor
-@AllArgsConstructor
-@RequiredArgsConstructor
-public class Student {
-
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    @Column(name = "id",updatable = false)
-    private Long id;
-
-    @Column(name = "name")
-    private String name;
-
-
-    @Column(name = "phone_number")
-    private String phone_number;
-
-    @Column(name = "image")
-    private String image;
-
-    @Column(name = "nickname")
-    private String nickname;
-
-
-}
