+++ backend/driving-today/src/main/java/com/drivingtoday/domain/academy/AcademyRepository.java
@@ -1,6 +1,5 @@
-package com.drivingtoday.repository;
+package com.drivingtoday.domain.academy;
 
-import com.drivingtoday.entity.Academy;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface AcademyRepository extends JpaRepository<Academy, Long> {
