+++ backend/driving-today/src/main/java/com/drivingtoday/domain/reservation/ReservationRepository.java
@@ -1,6 +1,5 @@
-package com.drivingtoday.repository;
+package com.drivingtoday.domain.reservation;
 
-import com.drivingtoday.entity.Reservation;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ReservationRepository extends JpaRepository<Reservation, Long> {
